refactor and extra logging for unpublished weekly actions
