+++ src/scheduler/workers/unpublish/findListsInAnnualReview.ts
@@ -1,7 +1,7 @@
 import { startOfToday } from "date-fns";
 import { prisma } from "server/models/db/prisma-client";
 
-export async function findListsInAnnualReviewForReminders() {
+export async function findListsInAnnualReview() {
   const today = startOfToday().toISOString();
   return await prisma.list.findMany({
     where: {
