+++ src/scheduler/workers/unpublish/types.ts
@@ -1,0 +1,6 @@
+import { List } from "@prisma/client";
+
+interface CountryName {
+  country: { name: string };
+}
+export type ListWithCountryName = List & CountryName;
