+++ src/scheduler/workers/unpublish/weekly/findNonRespondentsForList.ts
@@ -5,7 +5,7 @@ import { List, Prisma } from "@prisma/client";
 import { findReminderToSend } from "./findReminderToSend";
 
 export async function findNonRespondentsForList(list: List) {
-  const logger = schedulerLogger.child({ listId: list.id, method: "findNonRespondentsForList" });
+  const logger = schedulerLogger.child({ listId: list.id, method: "findNonRespondentsForList", timeframe: "weekly" });
 
   const jsonData = list.jsonData as ListJsonData;
   const { keyDates } = jsonData.currentAnnualReview!;
