+++ src/scheduler/workers/unpublish/weekly/getMetaForList.ts
@@ -1,7 +1,8 @@
-import { differenceInDays, differenceInWeeks, format, parseISO, startOfDay, startOfToday } from "date-fns";
+import { differenceInWeeks, format, parseISO, startOfDay, startOfToday } from "date-fns";
 import { List } from "server/models/types";
-import { logger } from "server/services/logger";
-import { ListWithCountryName, Meta } from "../types";
+import { ListWithCountryName } from "../types";
+import { Meta } from "./types";
+import { schedulerLogger } from "scheduler/logger";
 
 /**
  * {@link https://date-fns.org/v1.28.5/docs/format}
@@ -15,27 +16,24 @@ const DISPLAY_DATE_FORMAT = "d MMMM yyyy";
  * Additional data extracted from `List` to be passed down for each email.
  */
 export function getMetaForList(list: ListWithCountryName): Meta | undefined {
+  const logger = schedulerLogger.child({ listId: list.id, method: "getMetaForList", timeframe: "weekly" });
   const { jsonData } = list as List;
   const { currentAnnualReview } = jsonData;
 
   if (!currentAnnualReview) {
-    logger.error(
-      `getMetaForList: list.id ${list.id} does not have a fully qualified currentAnnualReview.keyDates object`
-    );
+    logger.error(`list ${list.id} does not have a fully qualified currentAnnualReview.keyDates object`);
     return;
   }
 
   const { keyDates } = currentAnnualReview;
 
-  const startDate = startOfDay(parseISO(keyDates.annualReview.START));
   const endDate = startOfDay(parseISO(keyDates.unpublished.UNPUBLISH));
+  const startDate = startOfDay(parseISO(keyDates.annualReview.START));
   const today = startOfToday();
-  const daysUntilUnpublish = differenceInDays(endDate, today);
 
   return {
     reference: jsonData.currentAnnualReview!.reference,
-    daysUntilUnpublish,
-    weeksUntilUnpublish: differenceInWeeks(endDate, startOfToday()),
+    weeksUntilUnpublish: differenceInWeeks(endDate, today),
     weeksSinceStart: differenceInWeeks(today, startDate, { roundingMethod: "floor" }),
     parsedUnpublishDate: format(endDate, DISPLAY_DATE_FORMAT),
     countryName: list.country.name,
