+++ src/scheduler/workers/unpublish/weekly/index.ts
@@ -1,9 +1,9 @@
 import { logger } from "server/services/logger";
 import { sendEmailsToNonRespondents } from "./sendEmailToNonRespondents";
-import { findListsInAnnualReviewForReminders } from "../findListsInAnnualReviewForReminders";
+import { findListsInAnnualReview } from "../findListsInAnnualReview";
 
 export async function main() {
-  const listsInAnnualReview = await findListsInAnnualReviewForReminders();
+  const listsInAnnualReview = await findListsInAnnualReview();
   const results = await Promise.allSettled(listsInAnnualReview.map(sendEmailsToNonRespondents));
   results
     .filter((result) => result.status !== "fulfilled")
