+++ src/scheduler/workers/unpublish/weekly/types.ts
@@ -1,5 +1,3 @@
-import { Address, List, ListItem } from "@prisma/client";
-
 export interface Meta {
   /**
    * annual review reference for list (`List.jsonData.currentAnnualReview.reference`)
@@ -10,9 +8,3 @@ export interface Meta {
   parsedUnpublishDate: string;
   countryName: string;
 }
-
-interface CountryName {
-  country: { name: string };
-}
-export type ListItemWithCountryName = ListItem & { address: Address & CountryName };
-export type ListWithCountryName = List & CountryName;
